% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.replace.na.R
\name{add_params}
\alias{add_params}
\title{add_params}
\usage{
add_params(ramclustObj, params, param_name)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{params}{vector containing parameters to add}

\item{param_name}{name of the parameter/step}
}
\value{
ramclustR object with rc.feature.replace.na params added.
}
\description{
add rc.feature.replace.na params in ramclustObj
}
