% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.my.names.R
\name{apply.my.names}
\alias{apply.my.names}
\title{apply.my.names}
\usage{
apply.my.names(ramclustObj = NULL, cmpds = NULL)
}
\arguments{
\item{ramclustObj}{a ramclustR object to name compounds}

\item{cmpds}{data.frame.  looking for column names 'cmpd.name', 'CID', and 'inchikey'.  one of 'CID' and 'inchikey' must be present.}
}
\value{
returns a ramclustObj with new annotation names based on your input.
}
\description{
replace DB names with names of your chosing, based on inchikey matching.
}
\details{
database chemical names can be a bit odd at times, and sometimes are just database reference numbers. This function takes output from 'get.inchikey.from.names' and uses it to rename compounds based on inchikey or pubchem CID matches.
}
\author{
Corey Broeckling
}
