% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.filter.blanks.R
\name{check_arguments_filter.blanks}
\alias{check_arguments_filter.blanks}
\title{check_arguments_filter.blanks}
\usage{
check_arguments_filter.blanks(ramclustObj, sn)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{sn}{numeric defines the ratio for 'signal'.  i.e. sn = 3 indicates that signal intensity must be 3 fold higher in sample than in blanks, on average, to be retained.}
}
\description{
check provided arguments
}
