% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.filter.cv.R
\name{check_arguments_filter.cv}
\alias{check_arguments_filter.cv}
\title{check_arguments_filter.cv}
\usage{
check_arguments_filter.cv(ramclustObj, qc.tag)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{qc.tag}{character vector of length one or two.  If length is two, enter search string and factor name in $phenoData slot (i.e. c("QC", "sample.type"). If length one (i.e. "QC"), will search for this string in the 'sample.names' slot by default.}
}
\description{
check provided arguments
}
