% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.replace.na.R
\name{check_arguments_replace.na}
\alias{check_arguments_replace.na}
\title{check_arguments_replace.na}
\usage{
check_arguments_replace.na(
  ramclustObj,
  replace.int,
  replace.noise,
  replace.zero
)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{replace.int}{default = 0.1.  proportion of minimum feature value to replace NA (or zero) values with}

\item{replace.noise}{default = 0.1.  proportion ofreplace.int value by which noise is added via 'jitter'}

\item{replace.zero}{logical if TRUE, any zero values are replaced with noise as if they were NA values}
}
\description{
check provided arguments
}
