% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.get.df.data.R
\name{checks}
\alias{checks}
\title{checks}
\usage{
checks(
  ms1_featureDefinitions = NULL,
  ms1_featureValues = NULL,
  ms2_featureValues = NULL,
  feature_names = NULL
)
}
\arguments{
\item{ms1_featureDefinitions}{dataframe with metadata with columns: mz, rt, feature names containing MS data}

\item{ms1_featureValues}{dataframe with rownames = sample names, colnames = feature names containing MS data}

\item{ms2_featureValues}{dataframe with rownames = sample names, colnames = feature names containing MSMS data}

\item{feature_names}{feature names extracted from the data}
}
\description{
check if MS data contains mz and rt, and if MSMS data is present feature names and sample names are identical
}
