% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpd.summary.R
\name{cmpd.summary}
\alias{cmpd.summary}
\title{cmpd.summary}
\usage{
cmpd.summary(ramclustObj = NULL, cmpd = 1)
}
\arguments{
\item{ramclustObj}{ramclustR object to annotate}

\item{cmpd}{integer. compound number to report.  i.e. 459.}
}
\description{
a bit of reporting for compounds, quick access summary and plot (if available)
}
\details{
Reports name, annotation, retention time, number of features in spectrum, median and mean signal intensity, and if interpretMSSpectrum (do.findmain) has been run, plots an annotated MS level spectrum.
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.

Broeckling CD, Ganna A, Layer M, Brown K, Sutton B, Ingelsson E, Peers G, Prenni JE. Enabling Efficient and Confident Annotation of LC-MS Metabolomics Data through MS1 Spectrum and Time Prediction. Anal Chem. 2016 Sep 20;88(18):9226-34. doi: 10.1021/acs.analchem.6b02479. Epub 2016 Sep 8. PubMed PMID: 7560453.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramlclustR}
