% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.filter.cv.R
\name{compute_do.sets}
\alias{compute_do.sets}
\title{compute_do.sets}
\usage{
compute_do.sets(ramclustObj)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}
}
\value{
vector which is used to select data frame to use in ramclustObj
}
\description{
compute data frame to use in ramclustObj
}
