% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramclustR.R
\name{compute_SpecAbundAve}
\alias{compute_SpecAbundAve}
\title{compute_SpecAbundAve}
\usage{
compute_SpecAbundAve(ramclustObj = NULL)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}
}
\value{
ramclustR object with aggregate by sample names for 'SpecAbundAve' dataset
}
\description{
further aggregate by sample names for 'SpecAbundAve' dataset
}
