% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.get.csv.data.R
\name{compute_wt_mean}
\alias{compute_wt_mean}
\title{compute_wt_mean}
\usage{
compute_wt_mean(data, global.min, fmz, ensure.no.na)
}
\arguments{
\item{data}{feature in ms/msms level data}

\item{global.min}{minimum intensity in ms/msms level data}

\item{fmz}{feature retention time}

\item{ensure.no.na}{logical: if TRUE, any 'NA' values in msint and/or msmsint are replaced with numerical values based on 10 percent of feature min plus noise.  Used to ensure that spectra are not written with NA values.}
}
\value{
weighted.mean intensity of feature in ms/msms level data
}
\description{
compute weighted.mean intensity of feature in ms/msms level data
}
