% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.get.df.data.R
\name{create_ramclustObj}
\alias{create_ramclustObj}
\title{create_ramclustObj}
\usage{
create_ramclustObj(
  ExpDes = NULL,
  input_history = NULL,
  MSdata = NULL,
  MSMSdata = NULL,
  frt = NULL,
  fmz = NULL,
  st = NULL,
  phenoData = NULL,
  feature_names = NULL,
  sample_names = NULL,
  xcmsOrd = NULL,
  ensure.no.na = TRUE
)
}
\arguments{
\item{ExpDes}{either an R object created by R ExpDes object: data used for record keeping and labelling msp spectral output}

\item{input_history}{input history}

\item{MSdata}{dataframe containing MS Data}

\item{MSMSdata}{dataframe containing MSMS Data}

\item{frt}{feature retention time, in whatever units were fed in}

\item{fmz}{feature retention time}

\item{st}{numeric: sigma t - time similarity decay value}

\item{phenoData}{dataframe containing phenoData}

\item{feature_names}{feature names extracted from the data}

\item{sample_names}{sample names extracted from the data}

\item{xcmsOrd}{original xcms order of features, for back-referencing when necessary}

\item{ensure.no.na}{logical: if TRUE, any 'NA' values in msint and/or msmsint are replaced with numerical values based on 10 percent of feature min plus noise.  Used to ensure that spectra are not written with NA values.}
}
\value{
an ramclustR object. this object is formatted as an hclust object with additional slots for holding feature and compound data.
}
\description{
create ramclustr Object
}
