% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Params.R
\name{defineExperiment}
\alias{defineExperiment}
\title{defineExperiment}
\usage{
defineExperiment(csv = TRUE)
}
\arguments{
\item{csv}{logical or filepath.  If   csv = TRUE , csv template called "ExpDes.csv" will be written to your working directory.  you will fill this in manually, ensuring that when you save you retain csv format.  ramclustR will then read this file in and and format appropriately.  If csv = FALSE, a pop up window will appear (in windows, at leaset) asking for input.  If a character string with full path (and file name) to a csv file is given, this will allow you to read in a previously edited csv file.}
}
\value{
an Exp Des R object which will be used for record keeping and writing spectra data.
}
\description{
Create an Experimental Design R object for record-keeping and msp output
}
\author{
Corey Broeckling
}
