% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.msfinder.formula.R
\name{do.msfinder.formula}
\alias{do.msfinder.formula}
\title{do.msfinder.formula}
\usage{

  do.msfinder.formula(msfinder.dir = "K:/software/MSFinder/MS-FINDER program ver. 2.20",
  mat.dir = NULL, msp.dir = NULL)
}
\arguments{
\item{msfinder.dir}{full path to the directory containing the MSFinderConsoleApp exe file}

\item{mat.dir}{by default, ramclustR will look in the working directory for a spectra/mat or spectra/msp subdirectory.  Else specify the path here for mat formatted spectra (all spectra in directory will be used)}

\item{msp.idr}{by default, ramclustR will look in the working directory for a spectra/mat or spectra/msp subdirectory.  Else specify the path here for msp formatted spectra (all spectra in directory will be used)}
}
\value{
nothing is returned.  data can be imported using the import.MSFinder.formulas() and import.MSFinder.formulas.R(structure) commands after processing has finished
}
\description{
Call MSFinder from R.  Currently requires clunky pasting into cmd prompt.
}
\details{
calls the 'MsfinderConsoleApp.exe' 'predict' program.  Ensure before calling that you have set appropriate parameters in the 'MSFINDER.INI' file
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.

Broeckling CD, Ganna A, Layer M, Brown K, Sutton B, Ingelsson E, Peers G, Prenni JE. Enabling Efficient and Confident Annotation of LC-MS Metabolomics Data through MS1 Spectrum and Time Prediction. Anal Chem. 2016 Sep 20;88(18):9226-34. doi: 10.1021/acs.analchem.6b02479. Epub 2016 Sep 8. PubMed PMID: 7560453.

Tsugawa H, Kind T, Nakabayashi R, Yukihira D, Tanaka W, Cajka T, Saito K, Fiehn O, Arita M. Hydrogen Rearrangement Rules: Computational MS/MS Fragmentation and Structure Elucidation Using MS-FINDER Software. Anal Chem. 2016 Aug 16;88(16):7946-58. doi: 10.1021/acs.analchem.6b00770. Epub 2016 Aug 4. PubMed PMID: 27419259.
}
\author{
Corey Broeckling
}
\keyword{'RAMClustR',}
\keyword{'clustering',}
\keyword{'feature',}
\keyword{'mass}
\keyword{'metabolomics',}
\keyword{'ramclustR'}
\keyword{'ramclustR',}
\keyword{'xcms'}
\keyword{spectrometry',}
