% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.filter.cv.R
\name{filter_good_features}
\alias{filter_good_features}
\title{filter_good_features}
\usage{
filter_good_features(ramclustObj, keep)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{keep}{features to keep. output of find_good_features().}
}
\value{
ramclustR object filtered to keep only 'good' features
}
\description{
filter to keep only 'good' features
}
