% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.filter.blanks.R
\name{filter_signal}
\alias{filter_signal}
\title{filter_signal}
\usage{
filter_signal(ms.qc.mean, ms.blank.mean, sn)
}
\arguments{
\item{ms.qc.mean}{ms qc mean signal intensities}

\item{ms.blank.mean}{ms blank mean signal intensities}

\item{sn}{numeric defines the ratio for 'signal'.  i.e. sn = 3 indicates that signal intensity must be 3 fold higher in sample than in blanks, on average, to be retained.}
}
\value{
union of which signal is at least 3x larger
}
\description{
filter signal
}
