% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.filter.cv.R
\name{find_good_features}
\alias{find_good_features}
\title{find_good_features}
\usage{
find_good_features(ramclustObj, do.sets, max.cv, qc)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{do.sets}{select data frame to use.}

\item{max.cv}{numeric maximum allowable cv for any feature.  default = 0.5}

\item{qc}{QC samples found by define_samples}
}
\value{
ramclustR object

features to keep
}
\description{
find 'good' features, acceptable CV at either MS or MSMS level results in keeping
}
