% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.inchikey.from.name.R
\name{get.inchikey.from.name}
\alias{get.inchikey.from.name}
\title{get.inchikey.from.name}
\usage{
get.inchikey.from.name(cmpd.names = NULL, citation.weights = FALSE)
}
\arguments{
\item{cmpd.names}{character vector.  i.e. c("caffeine", "theobromine", "glucose")}
}
\value{
returns a data.frame with three columns: name (supplied in cmpd.names), CID (pubchem cid), and inchikey. one row for each compound, compounds are processed in order supplied.
}
\description{
use rpubchem and Chemmimer tools to retreive inchikey from chemical names.
}
\details{
useful for moving from chemical name to digital structure represtation. greek letters are assumed to be 'UTF-8' encoded, and are converted to latin text before searching.   if you are reading in your compound name list, do so with 'encoding' set to 'UTF-8'.
}
\author{
Corey Broeckling
}
