% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pubchem.data.R
\name{get.pubchem.data}
\alias{get.pubchem.data}
\title{get.pubchem.data}
\usage{
get.pubchem.data(
  cmpd.names = NULL,
  cmpd.cid = NULL,
  cmpd.inchikey = NULL,
  use.parent.cid = TRUE,
  manual.entry = FALSE,
  get.vendors = TRUE,
  priority.vendors = c("Sigma Aldrich", "Alfa Chemistry", "Acros Organics", "VWR",
    "Alfa Aesar", "molport", "Key Organics", "BLD Pharm"),
  get.properties = TRUE,
  all.props = TRUE,
  get.bioassays = TRUE
)
}
\arguments{
\item{cmpd.names}{character vector.  i.e. c("caffeine", "theobromine", "glucose")}

\item{cmpd.cid}{numeric integer vector.  i.e. c(2519, 5429, 107526)}

\item{cmpd.inchikey}{character vector.  i.e. c("RYYVLZVUVIJVGH-UHFFFAOYSA-N", "YAPQBXQYLJRXSA-UHFFFAOYSA-N", "GZCGUPFRVQAUEE-SLPGGIOYSA-N")}

\item{use.parent.cid}{logical.  If TRUE, the CID for each supplied name/inchikey is used to retreive its parent CID (i.e. the parent of sodium palmitate is palmitic acid).  The parent CID is used to retrieve all other names, properties.}

\item{manual.entry}{logical.  if TRUE, user input is enabled for compounds not matched by name. A browser window will open with the pubchem search results in your default browser.}

\item{get.vendors}{logical.  if TRUE, vendor data is returned for each compound with a matched CID.  Includes vendor count and vendor product URL, if available}

\item{priority.vendors}{charachter vector.  i.e. c("MyFavoriteCompany", "MySecondFavoriteCompany").  If these vendors are found, the URL returned is from priority vendors. Priority is given by order input by user.}

\item{get.properties}{logical.  if TRUE, physicochemical property data are returned for each compound with a matched CID.}

\item{all.props}{logical.  If TRUE, all pubchem properties (https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865567) are returned.  If false, only a subset (faster).}

\item{get.bioassays}{logical. If TRUE, return a table summarizing existing bioassay data for that CID.}
}
\value{
returns a list with one or more of $puchem (compound name and identifiers) - one row in dataframe per CID; $properties contains pysicochemical properties - one row in dataframe per CID; $vendors contains the number of vendors for a given compound and selects a vendor based on 'priortity.vendors' supplied, or randomly choses a vendor with a HTML link - one row in dataframe per CID;  $bioassays contains a summary of bioassay activity data from pubchem - zero to many rows in dataframe per CID
}
\description{
use pubchem rest and view APIs to retreive structures, CIDs (if a name or inchikey is given), and optionally vendor data, when available.
}
\details{
useful for moving from chemical name to digital structure represtation. greek letters are assumed to be 'UTF-8' encoded, and are converted to latin text before searching.   if you are reading in your compound name list, do so with 'encoding' set to 'UTF-8'.
}
\author{
Corey Broeckling
}
