% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.synonyms.R
\name{get.synonyms}
\alias{get.synonyms}
\title{get.synonyms()}
\usage{
get.synonyms(ramclustObj = RC, delay.time = 0.5)
}
\arguments{
\item{ramclustObj}{R object - the ramclustR object which was used to write the .mat or .msp files}
}
\value{
an updated ramclustR object, with the RC$synonyms slot containing a list with length equal to the number of compounds in the dataset.  Each list element represents a character vector of all the synonyms returns (or NA, for compounds with no inchikey)
}
\description{
Use chemical translation service to retreive synonyms for ramclustR inchikeys
}
\details{
this function uses the chemical translation service (http://cts.fiehnlab.ucdavis.edu/) to look up known names (synonyms for a given inchikey)
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.

http://cts.fiehnlab.ucdavis.edu/static/download/CTS2-MS2015.pdf
}
\author{
Corey Broeckling
}
\keyword{'MSFinder',}
\keyword{'RAMClustR',}
\keyword{'chemical}
\keyword{'clustering',}
\keyword{'cts'}
\keyword{'feature',}
\keyword{'mass}
\keyword{'metabolomics',}
\keyword{'ramclustR'}
\keyword{'ramclustR',}
\keyword{'xcms',}
\keyword{service',}
\keyword{spectrometry',}
\keyword{translation}
