% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineExperiment.R
\name{get_ExpDes}
\alias{get_ExpDes}
\title{get_ExpDes}
\usage{
get_ExpDes(csv.in)
}
\arguments{
\item{csv.in}{Experimental Design read from csv}
}
\value{
list containing design and instrument
}
\description{
get Experimental Design
}
