% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineExperiment.R
\name{get_instrument_platform}
\alias{get_instrument_platform}
\title{get_instrument_platform}
\usage{
get_instrument_platform(design)
}
\arguments{
\item{design}{data frame containing Experimental Design}
}
\value{
instrument platform
}
\description{
get instrument platform
}
