% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClassyFire.R
\name{getClassyFire}
\alias{getClassyFire}
\title{getClassyFire}
\usage{
getClassyFire(ramclustObj = NULL, get.all = TRUE, max.wait = 10)
}
\arguments{
\item{ramclustObj}{ramclustR object to ClassyFy}

\item{get.all}{logical; if TRUE, when inchikey classyfire lookup fails, submits for classyfication.  Can be slow. max.wait (below) sets max time to spend on each compound before moving on. default = FALSE.}

\item{max.wait}{numeric; maximum time to wait per compound when 'get.all' = TRUE.  The}
}
\value{
returns a ramclustR object.  new dataframe in $classyfire slot with rows equal to number of compounds.
}
\description{
use classyfire web API to look up full ClassyFire heirarchy for each inchikey
}
\details{
The $inchikey slot is used to look up the
}
\references{
Djoumbou Feunang Y, Eisner R, Knox C, Chepelev L, Hastings J, Owen G, Fahy E, Steinbeck C, Subramanian S, Bolton E, Greiner R, and Wishart DS. ClassyFire: Automated Chemical Classification With A Comprehensive, Computable Taxonomy. Journal of Cheminformatics, 2016, 8:61. DOI: 10.1186/s13321-016-0174-y
}
\author{
Corey Broeckling
}
\keyword{'RAMClustR',}
\keyword{'classyFire'}
\keyword{'metabolomics',}
\keyword{'ramclustR'}
\keyword{'ramclustR',}
