% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(
  ramclustObj = RC,
  which.data = "SpecAbund",
  delim = "-",
  cmpdlabel = "cmpd",
  filter = FALSE
)
}
\arguments{
\item{ramclustObj}{ramclustR object to retrieve data from}

\item{which.data}{character; which dataset (SpecAbund or SpecAbundAve) to reference}

\item{delim}{character; "-" by default - the delimiter for parsing sample names to factors}

\item{cmpdlabel}{= "cmpd";  label the data with the annotation. can also be set to 'ann' for column names assigned as annotatins.}

\item{filter}{= TRUE; logical, if TRUE, checks for $cmpd.use slot generated by rc.cmpd.cv.filter() function, and only gets acceptable compounds.}
}
\value{
returns a list of length 3: $design is the experimental sample factors after parsing by the delim, $data is the dataset, $full.data is merged $des and $data data.frames.
}
\description{
retreive and parse sample names, retrieve metabolite data.  returns as list of two data frames
}
\details{
convenience function for parsing sample names and returning a dataset.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
