% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.get.xmcs.data.R
\name{getSampleMetadata}
\alias{getSampleMetadata}
\title{Extract Sample Metadata from an xcms Object}
\usage{
getSampleMetadata(xcmsObj)
}
\arguments{
\item{xcmsObj}{An xcms object containing chromatographic and metadata information.
This object is typically generated during preprocessing of 
metabolomics data using the xcms package.}
}
\value{
A data frame or list containing the sample metadata extracted from the xcms object.
        The structure and content of the returned metadata depend on the input xcms object.
}
\description{
This function retrieves sample metadata from a provided xcms object.
}
\details{
This function is designed to facilitate the extraction of metadata
         associated with samples in an xcms object. The metadata may include
         information such as sample names, injection order, batch information,
         and other experimental details.
}
\concept{xcms}
