% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.adap.kdb.R
\name{import.adap.kdb}
\alias{import.adap.kdb}
\title{import.adap.kdb}
\usage{
import.adap.kdb(
  ramclustObj = NULL,
  annotations = NULL,
  min.score = 700,
  annotate = TRUE,
  manual.name = TRUE
)
}
\arguments{
\item{ramclustObj}{ramclustR object to be annotated.}

\item{annotations}{file name/path to annotations .xlsx file.  generally 'simple_export.xlsx'}

\item{min.score}{700 (out of 1000) by default}

\item{annotate}{logical.  TRUE by default.  for now please leave default}

\item{manual.name}{when looking up inchikey/names, should manual input be used to fill ambiguous names? generally recommend TRUE}
}
\value{
returns a ramclustR structured object suitable for down stream processing steps.
}
\description{
use pubchem rest and view APIs to retrieve structures, CIDs (if a name or inchikey is given), synonyms, and optionally vendor data, when available.
}
\details{
useful for moving from chemical name to digital structure representation. greek letters are assumed to be 'UTF-8' encoded, and are converted to latin text before searching.   if you are reading in your compound name list, do so with 'encoding' set to 'UTF-8'.
}
\author{
Corey Broeckling
}
