% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.MSFinder.formulas.R
\name{import.MSFinder.formulas}
\alias{import.MSFinder.formulas}
\title{import.MSFinder.formulas}
\usage{
import.MSFinder.formulas(ramclustObj = RC, mat.dir = NULL,
  rt.form.filter = TRUE)
}
\description{
After running MSFinder on .mat or .msp files, import the formulas and scores that were predicted
}
