% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.filter.blanks.R
\name{mean_signal_intensities}
\alias{mean_signal_intensities}
\title{mean_signal_intensities}
\usage{
mean_signal_intensities(data, sample)
}
\arguments{
\item{data}{MS/MSMS data}

\item{sample}{sample found using the tag, output of define_samples()}
}
\value{
mean signal intensities
}
\description{
calculate MS mean signal intensities
}
