% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.normalize.tic.R
\name{normalized_data_tic}
\alias{normalized_data_tic}
\title{normalized_data_tic}
\usage{
normalized_data_tic(ramclustObj = NULL)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}
}
\value{
ramclustR object with total extracted ion normalized data.
}
\description{
normalize data using TIC
}
