% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.normalize.batch.qc.R
\name{order_datasets}
\alias{order_datasets}
\title{order_datasets}
\usage{
order_datasets(order = NULL, batch = NULL, qc = NULL, data = NULL)
}
\arguments{
\item{order}{integer vector with length equal to number of injections in xset or csv file or dataframe}

\item{batch}{integer vector with length equal to number of injections in xset or csv file or dataframe}

\item{qc}{logical vector with length equal to number of injections in xset or csv file or dataframe}

\item{data}{feature in ms/msms level data}
}
\value{
ordered feature in ms/msms level data, order, batch, qc
}
\description{
order the datasets first by batch and run order
}
