% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspecPDF.R
\name{plotspecPDF}
\alias{plotspecPDF}
\title{ramclustR}
\usage{
plotspecPDF(ramclustObj = RC, spec = NULL,
  mslev = as.numeric(as.character(RC$ExpDes[[2]]["MSlevs", 1])),
  mzdiff = 2.5)
}
\arguments{
\item{spec=NULL, }{}

\item{mslev=as.numeric(as.character(RC$ExpDes[[2]]["MSlevs", }{1])),}

\item{mzdiff=2.5}{}
}
\description{
Main clustering function
}
\details{
This is the Details section
}
\author{
Corey Broeckling
}

