\name{RAMClustR-package}
\alias{RAMClustR-package}
\alias{RAMClustR}
\docType{package}
\title{
RAMClustR: Feature clustering for metabolomics data enabling spectral matching
}
\description{
RAMClustR generates custom similiarity scores between metabolomics features, with XCMS output or properly formatted data matrices available as input.  Output consists of a new ramclust object, with feature cluster membership, a reduced dimensionality dataset, and msp formatted spectra file suitable for database searching.

}
\details{
\tabular{ll}{
Package: \tab RAMClustR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03-18\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Corey D. Broeckling, Fayyaz ul Amir Afsar Minhas, Steffen Neumann,
Asa Ben-Hur, Jessica E. Prenni

Maintainer: "Broeckling,Corey" <Corey.Broeckling@ColoState.EDU>, "Steffen Neumann" <sneumann@ipb-halle.de>,
}
\references{
Broeckling CD, Heuberger AL, Prince JA, Ingelsson E, Prenni JE (2013) Assigning precursor-product ion relationships in indiscriminant MS/MS data from non-targeted metabolite profiling studies. Metabolomics 9: 33-43. doi: 10.1038/nrg2344
}
\keyword{ metabolomics mass-spectrometry clustering}
\seealso{
\code{\link[<pkg>:xcms-package]{xcms}} \code{\link[<pkg>:CAMERA-package]{CAMERA}}
}
