\name{ramclustR}
\alias{ramclustR}
\title{ramclustR}
\usage{
  ramclustR (xcmsObj = NULL, ms = NULL, idmsms = NULL, taglocation = "filepaths", 
    MStag = NULL, idMSMStag = NULL, featdelim = "_", timepos = 2, 
    mzpos = 1, st = NULL, sr = NULL, maxt = NULL, deepSplit = FALSE, 
    blocksize = 2000, mult = 5, hmax = NULL, sampNameCol = NULL, 
    collapse = TRUE, mspout = TRUE, mslev = 2, ExpDes = NULL, 
    normalize = "TIC", minModuleSize = 2)
}
\arguments{
  \item{xcmsObj}{character Filename of the nmrML to check}

  \item{ms}{MS1 intensities =MSdata,}

  \item{idmsms}{=ms,}
    
  \item{taglocation}{="filepaths", either "filepaths" or "phenoData".  THe location for the tag in the xcms object in which to look for the idMSMStag (see below)}

  \item{idMSMStag}{character e.g. "02.cdf".  THe character string in the taglocation by which to tag files as indiscriminant MS/MS data}

  \item{featdelim}{character e.g. ="_".  If importing data from another peak detection tool, feature names must contain mass and retention time.  The delimiter separating them can is defined here}

  \item{timepos}{numeric 2. If importing data from another peak detection tool, feature names must contain mass and retention time. The position of the time component of the feature name (either 1 or 2) is defined here.}

  \item{st}{numeric e.g. =5,  Sigma t: When an xcms object is used as input, this parameter is by default set to half the peak width of all the peaks in the xcmsSet.  Can be manually overridden to tune performance.  Larger values are less selective, smaller values are more selective.}

  \item{sr}{numeric e.g. 0.5, Sigma r: the rate of similarity decay in response to correlational r between a pair of features.  the default values of this paratmer is the 0.5.  Can be manually overridden to tune performance.  Larger values are less selective, smaller values are more selective. }

  \item{maxt}{numeric e.g 20, the maximum time allowed for similarity between features to be greater than zero.  For example, if two features are separated by more than 20 seconds, the similarity score is automatically set to zero, ensuring that they end up in different clusters.  If you are importing data from another vendor, ensure that either you change the feature times to seconds (default in ramclustR) or you change this value to a value representing minutes, rather than seconds (e.g. 0.33)}

  \item{deepSplit}{boolean e.g. =FALSE,  access to the deepSplit function in dynamicTreeCut}

  \item{blocksize}{integer number of features processed in one block =2000, ramclustR makes attempts to minimize memory demands and maintain processing speed.  It does so by processing the full similarity matrix in blocks of blocksize dimensions.  This has two benefits: one is that a given PC is less likely to run out of memory during processing and the second is that we need not process blocks that contain not feature pairs with a delta rt of greater than maxt, thereby avoiding needless calculations and processing time.}

  \item{mult}{numeric =10, internal handler for reshaping blocksize}
}
\value{
  a RAMClustR object. THe base object structure is that of an hclust object, thus the plot methods available for other R objects of hlcust format are available.  In addition, several slots are added:
  featclus: the cluster to which the feature belongs. zero values indicate singletons. 
  frt: feature retention time used in clustering
  fmz: feature m/z value used to draw spectrum
  nfeat: the number of features contained in each cluster
  nsing: the number of singletons following clustering
  SpecAbund: the reduced dataset.  Features within a cluster are aggregated using a weighted mean function into a single value for each file
  SpecAbundAve: an (optionally) further reduced dataset in which the injection replicates have been averaged.
  ExpDes: the instrument parameters used for generation of raw data, the values of which are written to the output msp file. 
  MSdata: a matrix version of the low collision energy feature intensities
  MSMSdata: a matrix version of the high collision energy intensities
  cmpd: acluster name: autogenerated as C1... Cn, where n is the number of clusters with at least minModuleSize features
  ann: a place holder for compound annotations.  by default, all clusters are annotated with the label C1... Cn, where n is the number of clusters with at least minModuleSize features
  annconf: a placeholder for a numerical annotation confidence for each annotation (see Metabolomics Standards initiative for more information)
  annnotes: a placeholder for a vector of annotation notes: in the event of an ambiguous annotation, notes can be used to document the evidence and/or caveats
}
\description{
  Main clustering function
}
\details{
  This is the Details section
}
\author{
  Corey Broeckling
}

