% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramclustR2.R
\name{ramclustR2}
\alias{ramclustR2}
\title{ramclustR2}
\usage{
ramclustR2(xcmsObj = NULL, ms = NULL, idmsms = NULL,
  taglocation = "filepaths", MStag = NULL, idMSMStag = NULL,
  featdelim = "_", timepos = 2, pkshape = FALSE, sr = NULL, ss = NULL,
  st = NULL, maxt = NULL, deepSplit = FALSE, blocksize = 2000,
  mult = 5, hmax = NULL, sampNameCol = NULL, collapse = TRUE,
  usePheno = TRUE, mspout = TRUE, mslev = 1, ExpDes = NULL,
  normalize = "TIC", minModuleSize = 2, linkage = "average", mzdec = 4,
  saveFF = TRUE, plots = TRUE, cor.use = "everything",
  method = "spearman", min.scans = NULL)
}
\arguments{
\item{ms}{MS1 intensities =MSdata,}

\item{idmsms}{=ms,}

\item{idMSMStag}{character e.g. "02.cdf"}

\item{featdelim}{character e.g. ="_"}

\item{timepos}{numeric 2}

\item{sr}{numeric also no clue yet = 5,}

\item{st}{numeric no clue e.g. = 5,}

\item{maxt}{numeric again no clue =20,}

\item{deepSplit}{boolean e.g. =FALSE,}

\item{blocksize}{integer number of features (scans?) processed in one block  =1000,}

\item{mult}{numeric =10}

\item{filename}{character Filename of the nmrML to check}
}
\value{
A vector with the numeric values of the processed data
}
\description{
Main clustering function
}
\details{
This is the Details section
}
\author{
Corey Broeckling
}
