% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.export.msp.rc.R
\name{rc.export.msp.rc}
\alias{rc.export.msp.rc}
\title{rc.export.msp.rc}
\usage{
rc.export.msp.rc(ramclustObj = NULL, one.file = TRUE, mzdec = 1)
}
\arguments{
\item{ramclustObj}{ramclustR object to annotate.}

\item{one.file}{logical, should all msp spectra be written to one file? If false, each spectrum is an individual file.}

\item{mzdec}{integer.  Number of decimal points to export mass values with.}
}
\value{
nothing, just exports files to the working directory
}
\description{
Cluster annotation function: inference of 'M' - molecular weight of the compound giving rise to each spectrum - using the InterpretMSSpectrum::findMain function
}
\details{
exports files to a directory called 'spectra'.  If one.file = FALSE, a new directory 'spectra/msp' is created to hold the individual msp files. if do.findman has been run, spectra are written as ms2 spectra, else as ms1.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
\concept{clustering}
\concept{interpretMSSpectrum}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
