% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.cv.filter.R,
%   R/rc.feature.filter.cv.R
\name{rc.feature.filter.cv}
\alias{rc.feature.filter.cv}
\title{rc.feature.filter.cv}
\usage{
rc.feature.filter.cv(ramclustObj = NULL, qc.tag = "QC", max.cv = 0.5)

rc.feature.filter.cv(ramclustObj = NULL, qc.tag = "QC", max.cv = 0.5)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{qc.tag}{character vector of length one or two.  If length is two, enter search string and factor name in $phenoData slot (i.e. c("QC", "sample.type"). If length one (i.e. "QC"), will search for this string in the 'sample.names' slot by default.}

\item{max.cv}{numeric maximum allowable cv for any feature.  default = 0.3}
}
\value{
ramclustR object with total extracted ion normalized data.

ramclustR object with total extracted ion normalized data.
}
\description{
extractor for xcms objects in preparation for clustering

extractor for xcms objects in preparation for clustering
}
\details{
This function offers normalization by total extracted ion signal.  it is recommended to first run 'rc.feature.filter.blanks' to remove non-sample derived signal.

This function offers normalization by total extracted ion signal.  it is recommended to first run 'rc.feature.filter.blanks' to remove non-sample derived signal.
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.

Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.
}
\author{
Corey Broeckling

Corey Broeckling
}
\concept{MSFinder}
\concept{RAMClustR}
\concept{clustering}
\concept{feature}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
