% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.normalize.batch.qc.R
\name{rc.feature.normalize.batch.qc}
\alias{rc.feature.normalize.batch.qc}
\title{rc.feature.normalize.batch.qc}
\usage{
rc.feature.normalize.batch.qc(
  order = NULL,
  batch = NULL,
  qc = NULL,
  ramclustObj = NULL,
  qc.inj.range = 20,
  output.plot = FALSE
)
}
\arguments{
\item{order}{integer vector with length equal to number of injections in xset or csv file or dataframe}

\item{batch}{integer vector with length equal to number of injections in xset or csv file or dataframe}

\item{qc}{logical vector with length equal to number of injections in xset or csv file or dataframe}

\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{qc.inj.range}{integer: how many injections around each injection are to be scanned for presence of QC samples when using batch.qc normalization?  A good rule of thumb is between 1 and 3 times the typical injection span between QC injections.  i.e. if you inject QC ever 7 samples, set this to between 7 and 21.  smaller values provide more local precision but make normalization sensitive to individual poor outliers (though these are first removed using the boxplot function outlier detection), while wider values provide less local precision in normalization but better stability to individual peak areas.}

\item{output.plot}{logical set to TRUE to store plots}
}
\value{
ramclustR object with normalized data.
}
\description{
normalize data using batch.qc
}
