% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.normalize.quantile.R
\name{rc.feature.normalize.quantile}
\alias{rc.feature.normalize.quantile}
\title{rc.feature.normalize.quantile}
\usage{
rc.feature.normalize.quantile(ramclustObj = NULL)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}
}
\value{
ramclustR object with normalized data.
}
\description{
normalize data using quantile
}
