% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.qc.cv.R
\name{rc.feature.qc.cv}
\alias{rc.feature.qc.cv}
\title{rc.feature.qc.cv}
\usage{
rc.feature.qc.cv(ramclustObj = NULL, qc.tag = "QC", which.data = "SpecAbund")
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{qc.tag}{character vector of length one or two.  If length is two, enter search string and factor name in $phenoData slot (i.e. c("QC", "sample.type"). If length one (i.e. "QC"), will search for this string in the 'sample.names' slot by default.}

\item{which.data}{either 'MSdata' or 'MSMSdata' - the latter only if a DIA approach such as MSe has been utilized and apprpriately processed.}
}
\value{
ramclustR object with normalized data.
}
\description{
extractor for xcms objects in preparation for clustering
}
\details{
this function returns a vector of coefficiencts of varation - one for each feature, for QC samples recognized by qc.tag
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.
}
\author{
Corey Broeckling
}
\concept{MSFinder}
\concept{RAMClustR}
\concept{clustering}
\concept{feature}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
