% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.replace.na.mai.R
\name{rc.feature.replace.na.mai}
\alias{rc.feature.replace.na.mai}
\title{rc.feature.replace.na.mai}
\usage{
rc.feature.replace.na.mai(
  ramclustObj = NULL,
  MCAR_algorithm = "BPCA",
  MNAR_algorithm = "Single",
  forest_list_args = list(ntree = 300, proximity = FALSE),
  which.data = c("MSdata", "MSMSdata")
)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{MCAR_algorithm}{character.  default = "BPCA", see ?MAI::MAI}

\item{MNAR_algorithm}{character. default = "Single", see ?MAI::MAI}

\item{forest_list_args}{list. default = list(ntree = 300, proximity = FALSE), see ?MAI::MAI}

\item{which.data}{name of dataset.  default = c("MSdata", "MSMSdata")}
}
\value{
ramclustR object with NA and zero values removed.
}
\description{
replaces any NA (and optionally zero) values with small signal (20% of minimum feature signal value + 20% random noise)
}
\details{
noise is added by finding for each feature the minimum detected value, multiplying that value by replace.int, then adding (replace.int*replace.noise) noise.  abs() is used to ensure no negative values result.
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.
}
\author{
Corey Broeckling
}
\concept{MSFinder}
\concept{RAMClustR}
\concept{clustering}
\concept{feature}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
