% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.merge.clusters.R
\name{rc.merge.split.clusters}
\alias{rc.merge.split.clusters}
\title{rc.merge.split.clusters}
\usage{
rc.merge.split.clusters(
  ramclustObj = NULL,
  merge.threshold = 0.7,
  cor.method = "spearman",
  rt.sd.factor = 3,
  cor.use = "everything",
  sample.name.column = "sample.ID"
)
}
\arguments{
\item{ramclustObj}{ramclustR object to annotate.}

\item{merge.threshold}{numeric. value between -1 and 1 indicating the correlational r threshold above which two clusters will be merged}

\item{cor.method}{character: which correlational method used to calculate 'r' - see ?cor 'method' option. default = "pearson"}

\item{rt.sd.factor}{numeric.  default = 3.  clusters within rt.sd.factor * ramclustObj$rtsd (cluster retention time standard deviation) are considered for merging.}

\item{cor.use}{character: which data points to use to calculate 'r' - see ?cor 'use' option. default = "pairwise.complete.obs"}

\item{sample.name.column}{character. column name from ramclustObj$phenoData which should be used as row.names in new ramclustObj$SpecAbund dataset.}
}
\value{
new ramclustR object, with (generally) fewer clusters than the input ramclustR object.
}
\description{
Cluster refinement - scanning instruments (quadrupole, as in GC-MS) can display cluster splitting, possibily due to slight differences in measured peak retentiont time as a function of mass due to scan dynamics.  this function enables a second pass clustering designed to merge two clusters if the second cluster is within a small retention time window and shows a sufficiently strong correlation.
}
\details{
exports files to a directory called 'spectra'.  If one.file = FALSE, a new directory 'spectra/msp' is created to hold the individual msp files. if do.findman has been run, spectra are written as ms2 spectra, else as ms1.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
\concept{clustering}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
