% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.normalize.batch.qc.R
\name{rc.normalize.batch.qc}
\alias{rc.normalize.batch.qc}
\title{rc.normalize.batch.qc}
\usage{
rc.normalize.batch.qc(
  ramclustObj = NULL,
  qc.inj.range = 20,
  order = NULL,
  batch = NULL,
  qc = NULL
)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{qc.inj.range}{integer: how many injections around each injection are to be scanned for presence of QC samples when using batch.qc normalization?  A good rule of thumb is between 1 and 3 times the typical injection span between QC injections.  i.e. if you inject QC ever 7 samples, set this to between 7 and 21.  smaller values provide more local precision but make normalization sensitive to individual poor outliers (though these are first removed using the boxplot function outlier detection), while wider values provide less local precision in normalization but better stability to individual peak areas.}

\item{order}{integer vector with length equal to number of injections in xset or csv file}

\item{batch}{integer vector with length equal to number of injections in xset or csv file}

\item{qc}{logical vector with length equal to number of injections in xset or csv file.}

\item{normalize}{character: either "none", "TIC", "quantile", or "batch.qc" normalization of feature intensities.  see batch.qc overview in details.}
}
\value{
ramclustR object with normalized data.
}
\description{
extractor for xcms objects in preparation for clustering
}
\details{
This function offers normalization by run order, batch number, and QC sample signal intensity.

Each input vector should be the same length, and equal to the number of samples in the $MSdata set.

Input vector order is assumed to be the same as the sample order in the $MSdata set.
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.
}
\author{
Corey Broeckling
}
\concept{MSFinder}
\concept{RAMClustR}
\concept{clustering}
\concept{feature}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
