% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.normalize.feature.R
\name{rc.normalize.feature}
\alias{rc.normalize.feature}
\title{rc.normalize.feature}
\usage{
rc.normalize.feature(
  ramclustObj = NULL,
  mz = NULL,
  mztol = 0.01,
  rt = NULL,
  rttol = 3
)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{mz}{numerical vector - accurate mass(es) of feature to use for normalization}

\item{rt}{numerical vector same length as mz - retention times (in seconds, if xcms input) of feature to use for normalization}
}
\value{
ramclustR object with MSdata and MSMSdata normalized to selected features.
}
\description{
extractor for xcms objects in preparation for clustering
}
\details{
This function normalizes all dataset values relative to the summed intensity of all features selected using mz and rt vectors.
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.
}
\author{
Corey Broeckling
}
\concept{MSFinder}
\concept{RAMClustR}
\concept{clustering}
\concept{feature}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
