% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.restore.qc.samples.R
\name{rc.restore.qc.samples}
\alias{rc.restore.qc.samples}
\title{rc.restore.qc.samples}
\usage{
rc.restore.qc.samples(ramclustObj = NULL)
}
\arguments{
\item{ramclustObj}{ramclustR object to analyze}
}
\value{
RC object
}
\description{
summarize quality control for clustering and for quality control sample variation based on compound ($SpecAbund) and feature ($MSdata and $MSMSdata, if present)
}
\details{
moves all of $phenoData, $MSdata, $MSMSdata, $SpecAbund back to original positions from $qc slot
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.

Broeckling CD, Ganna A, Layer M, Brown K, Sutton B, Ingelsson E, Peers G, Prenni JE. Enabling Efficient and Confident Annotation of LC-MS Metabolomics Data through MS1 Spectrum and Time Prediction. Anal Chem. 2016 Sep 20;88(18):9226-34. doi: 10.1021/acs.analchem.6b02479. Epub 2016 Sep 8. PubMed PMID: 7560453.
}
\author{
Corey Broeckling
}
\concept{MSFinder}
\concept{RAMClustR}
\concept{clustering}
\concept{feature}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
