% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.filter.blanks.R
\name{remove_blanks}
\alias{remove_blanks}
\title{remove_blanks}
\usage{
remove_blanks(ramclustObj, blank)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{blank}{blank samples found by define_samples}
}
\value{
ramclustObj object with blanks removed
}
\description{
remove blanks
}
