% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.replace.na.R
\name{replace_na}
\alias{replace_na}
\title{replace_na}
\usage{
replace_na(data, replace.int, replace.zero, replace.noise)
}
\arguments{
\item{data}{selected data frame to use}

\item{replace.int}{default = 0.1.  proportion of minimum feature value to replace NA (or zero) values with}

\item{replace.zero}{logical if TRUE, any zero values are replaced with noise as if they were NA values}

\item{replace.noise}{default = 0.1.  proportion ofreplace.int value by which noise is added via 'jitter'}
}
\value{
selected ramclustR data frame with NA and zero values removed.

number of features replaced
}
\description{
add rc.feature.replace.na params in ramclustObj
}
