% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.single.msp.R
\name{write.single.msp}
\alias{write.single.msp}
\title{write.single.msp}
\usage{
write.single.msp(ramclustObj = NULL, mzdec = 4)
}
\arguments{
\item{ramclustObj}{your ramclustR output R object.}

\item{mzdec}{integer: number of decimal places used in printing m/z values}
}
\value{
- 'spectra/msp' directory is created in the working directory.  MS level 1 spectra are labelled with compound name, while MS2 spectra have an appended "MS2" in the filename.
}
\description{
writes single msp files compatible with MSfinder or other spectral interpretation tools.
}
\details{
writes individual msp files for each compound to a new 'spectra/msp' directory
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.

Broeckling CD, Ganna A, Layer M, Brown K, Sutton B, Ingelsson E, Peers G, Prenni JE. Enabling Efficient and Confident Annotation of LC-MS Metabolomics Data through MS1 Spectrum and Time Prediction. Anal Chem. 2016 Sep 20;88(18):9226-34. doi: 10.1021/acs.analchem.6b02479. Epub 2016 Sep 8. PubMed PMID: 7560453.
}
\author{
Corey Broeckling
}
\keyword{'RAMClustR',}
\keyword{'clustering',}
\keyword{'feature',}
\keyword{'mass}
\keyword{'metabolomics',}
\keyword{'ramclustR'}
\keyword{'ramclustR',}
\keyword{'xcms'}
\keyword{spectrometry',}
