% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineExperiment.R
\name{write_csv}
\alias{write_csv}
\title{write_csv}
\usage{
write_csv(data)
}
\arguments{
\item{data}{csv template to write}
}
\value{
read ExpDes.csv file
}
\description{
write csv template called "ExpDes.csv" to your working directory. you will fill this in manually, ensuring that when you save you retain csv format. ramclustR will then read this file in and and format appropriately.
}
