% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.summary.R
\name{annotation.summary}
\alias{annotation.summary}
\title{annotation.summary()}
\usage{
annotation.summary(ramclustObj = NULL, outfile = NULL)
}
\arguments{
\item{ramclustObj}{R object - the ramclustR object which was used to write the .mat or .msp files}

\item{outfile}{file path/name of output csv summary file.  if NULL (default) will be exported to spectra/annotaionSummary.csv}
}
\value{
nothing
}
\description{
Write a .csv file containing a summary of the annotations in the ramclustR object.
}
\details{
this function exports a csv file summarizing annotation evidence for each compound
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.
}
\author{
Corey Broeckling
}
\concept{MSFinder}
\concept{RAMClustR}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramlclustR}
\concept{xcms}
