% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.feature.filter.cv.R
\name{define_samples}
\alias{define_samples}
\title{define_samples}
\usage{
define_samples(ramclustObj, tag)
}
\arguments{
\item{ramclustObj}{ramclustObj containing MSdata with optional MSMSdata (MSe, DIA, idMSMS)}

\item{tag}{character vector of length one or two.  If length is two, enter search string and factor name in $phenoData slot (i.e. c("QC", "sample.type"). If length one (i.e. "QC"), will search for this string in the 'sample.names' slot by default.}
}
\value{
samples found using the tag
}
\description{
define samples in each set
}
