% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.findmain.R
\name{do.findmain}
\alias{do.findmain}
\title{do.findmain}
\usage{
do.findmain(ramclustObj = RC, cmpd = NULL, mode = "positive",
  mzabs.error = 0.01, ppm.error = 10, ads = NULL, nls = NULL,
  adwts = NULL, nlwts = NULL, plot.findmain = TRUE, writeMat = TRUE)
}
\arguments{
\item{ramclustObj}{ramclustR object to annotate.}

\item{cmpd}{integer: vector defining compound numbers to annotated.  if NULL (default), all compounds}

\item{mode}{character: "positive" or "negative"}

\item{mzabs.error}{numeric: absolute mass deviation allowd}

\item{ppm.error}{numeric: ppm mass error added to mzabs.error}

\item{ads}{character: vector of allowed adducts, i.e. c("[M+H]+")}

\item{nls}{character: vector of allowed neutral losses, i.e. c("[M+H-H2O]+")}

\item{adwts}{numeric: vector of weights for adducts. length should match that of 'ads', else first value will be repeated.  if NULL (default), value of 1 is assigned to all. Only used for ramclustR scoring of M.}

\item{nlwts}{numeric: vector of weights for neutral losses. length should match that of 'nls', else first value will be repeated.  if NULL (default), value of 0.1 is assigned to all.  Only used for ramclustR scoring of M.}

\item{plot.findmain}{logical: should pdf polts be generated for evaluation?}

\item{writeMat}{logical: should indidual .mat files (for MSFinder) be generated in a 'mat' subdirectory in the 'spectra' folder?}
}
\value{
an annotated ramclustR object.  base structure is that of a standard R heirarchical clustering output
}
\description{
cluster annotation function
}
\details{
This is the Details section
}
\author{
Corey Broeckling
}
