% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDataset.R
\name{exportDataset}
\alias{exportDataset}
\title{exportDataset}
\usage{
exportDataset(
  ramclustObj = NULL,
  which.data = "SpecAbund",
  label.by = "ann",
  appendFactors = TRUE
)
}
\arguments{
\item{ramclustObj}{ramclustR object to export from}

\item{which.data}{name of dataset to export.  SpecAbund, SpecAbundAve, MSdata, or MSMSdata}

\item{label.by}{either 'ann' or 'cmpd', generally.  name of ramclustObj slot used as csv header for each column (compound)}

\item{appendFactors}{logical.  If TRUE (default) the factor data frame is appended to the left side of the dataset.}
}
\value{
nothing is returned.  file exported as csf to 'datasets/*.csv'
}
\description{
export one of 'SpecAbund', 'SpecAbundAve', 'MSdata' or 'MSMSdata' from an RC object to csv
}
\details{
Useful for exporting the processed signal intensity matrix to csv for analysis elsewhere.
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
\concept{ramclustR}
