% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pubchem.data.R
\name{get.pubchem.data}
\alias{get.pubchem.data}
\title{get.pubchem.data}
\usage{
get.pubchem.data(
  cmpd.names = NULL,
  cmpd.cid = NULL,
  cmpd.inchikey = NULL,
  manual.entry = FALSE,
  vendor.data = TRUE,
  priority.vendors = c("Sigma Aldrich", "Alfa Chemistry", "Acros Organics", "VWR",
    "Alfa Aesar")
)
}
\arguments{
\item{cmpd.names}{character vector.  i.e. c("caffeine", "theobromine", "glucose")}

\item{cmpd.cid}{numeric integer vector.  i.e. c(2519, 5429, 107526)}

\item{cmpd.inchikey}{character vector.  i.e. c("RYYVLZVUVIJVGH-UHFFFAOYSA-N", "YAPQBXQYLJRXSA-UHFFFAOYSA-N", "GZCGUPFRVQAUEE-SLPGGIOYSA-N")}

\item{manual.entry}{logical.  if TRUE, user input is enabled for compounds not matched by name. A browser window will open with the pubchem search results in your default browser.}

\item{vendor.data}{logical.  if TRUE, vendor data is returned for each compound with a matched CID.  Includes vendor count and vendor product URL, if available}

\item{priority.vendors}{charachter vector.  i.e. c("MyFavoriteCompany", "MySecondFavoriteCompany").  If these vendors are found, the URL returned is from priority vendors. Priority is given by order input by user.}
}
\value{
returns a data.frame with several columns: "compound", "CID", "CanonicalSMILES", "IsomericSMILES", "InChI", "InChIKey", "pubchem.url", "pubchem.vendors.url", "n.vendors", "vendor.url"
}
\description{
use pubchem rest and view APIs to retreive structures, CIDs (if a name or inchikey is given), and optionally vendor data, when available.
}
\details{
useful for moving from chemical name to digital structure represtation. greek letters are assumed to be 'UTF-8' encoded, and are converted to latin text before searching.   if you are reading in your compound name list, do so with 'encoding' set to 'UTF-8'.
}
\author{
Corey Broeckling
}
