% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClassyFire.R
\name{getClassyFire}
\alias{getClassyFire}
\title{getClassyFire}
\usage{
getClassyFire(ramclustObj = RC)
}
\arguments{
\item{ramclustObj}{ramclustR object to ClassyFy}
}
\value{
returns a ramclustR object.  All values from ramclustObj.1 are retained.  SpecAbund dataset from ramclustObj.1 is moved to RC$SpecAbund.1, where RC is the new ramclustObj.
}
\description{
use classyfire web API to look up full ClassyFire heirarchy for each inchikey
}
\details{
The $inchikey slot is used to look up the
}
\references{
Djoumbou Feunang Y, Eisner R, Knox C, Chepelev L, Hastings J, Owen G, Fahy E, Steinbeck C, Subramanian S, Bolton E, Greiner R, and Wishart DS. ClassyFire: Automated Chemical Classification With A Comprehensive, Computable Taxonomy. Journal of Cheminformatics, 2016, 8:61. DOI: 10.1186/s13321-016-0174-y
}
\author{
Corey Broeckling
}
\keyword{'RAMClustR',}
\keyword{'classyFire'}
\keyword{'metabolomics',}
\keyword{'ramclustR'}
\keyword{'ramclustR',}
