% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.MSFinder.structures.R
\name{import.msfinder.structures}
\alias{import.msfinder.structures}
\title{write.methods}
\usage{
import.msfinder.structures(ramclustObj = NULL, mat.dir = NULL, msp.dir = NULL)
}
\arguments{
\item{ramclustObj}{R object - the ramclustR object which was used to write the .mat or .msp files}

\item{mat.dir}{directory in which to look for mat file MSFinder output - by default the /spectra/mat in the working directory}

\item{msp.dir}{directory in which to look for msp file MSFinder output - by default the /spectra/msp in the working directory}
}
\value{
an annotated ramclustR object

nothing - new file written to working director
}
\description{
write RAMClustR processing methods and citations to text file
}
\details{
this function exports a file called ramclustr_methods.txt which contains the processing history, parameters used, and relevent citations.
}
\references{
Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE. RAMClust: a novel feature clustering method enables spectral-matching-based annotation for metabolomics data. Anal Chem. 2014 Jul 15;86(14):6812-7. doi: 10.1021/ac501530d.  Epub 2014 Jun 26. PubMed PMID: 24927477.
}
\author{
Corey Broeckling
}
\concept{MSFinder}
\concept{RAMClustR}
\concept{clustering}
\concept{feature}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
