\name{RAMClustR-package}
\alias{RAMClustR-package}
\alias{RAMClustR}
\docType{package}
\title{
Assigning precursor-product ion relationships in indiscriminant MS/MS
data
}
\description{
Assigning precursor-product ion relationships in indiscriminant MS/MS
data
A concise (1-5 lines) description of the package ~~
}
\details{
\tabular{ll}{
Package: \tab RAMClustR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-10-23\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Corey D. Broeckling, Adam L. Heuberger, Jonathan A. Prince,
E. Ingelsson, Jessica E. Prenni

Maintainer: "Broeckling,Corey" <Corey.Broeckling@ColoState.EDU>, "Steffen Neumann" <sneumann@ipb-halle.de>,
}
\references{
Broeckling, C.D. Heuberger, A.L., Prince, J.A., Ingelsson, E., Prenni,
J.E., Assigning precursor-product ion relationships in indiscriminant
MS/MS data from non-targeted metabolite profiling studies. Metabolomics,
2012, In Press.
}
\keyword{ metabolomics massspectrometry clustering}
\seealso{
\code{\link[<pkg>:xcms-package]{xcms}} \code{\link[<pkg>:CAMERA-package]{CAMERA}}
}
