% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramclustR.R
\name{ramclustR}
\alias{ramclustR}
\title{ramclustR}
\usage{
ramclustR(xcmsObj = NULL, ms = NULL, idmsms = NULL,
  taglocation = "filepaths", MStag = NULL, idMSMStag = NULL,
  featdelim = "_", mzpos = 1, timepos = 2, st = NULL, sr = NULL,
  maxt = NULL, deepSplit = FALSE, blocksize = 2000, mult = 5,
  hmax = NULL, sampNameCol = NULL, collapse = TRUE, usePheno = TRUE,
  mspout = TRUE, ExpDes = NULL, normalize = "TIC", minModuleSize = 2,
  linkage = "average", mzdec = 4, cor.method = "pearson")
}
\arguments{
\item{xcmsObj}{xcmsObject: containing grouped feature data for clustering by ramclustR}

\item{ms}{filepath: optional csv input. Features as columns, rows as samples. Column header mz_rt}

\item{idmsms}{filepath: optional idMSMS / MSe csv data.  same dim and names as ms required}

\item{taglocation}{character: "filepaths" by default, "phenoData[,1]" is another option. referse to xcms slot}

\item{idMSMStag}{character: character string in 'taglocation' to designat idMSMS / MSe files e.g. "02.cdf"}

\item{featdelim}{character: how feature mz and rt are delimited in csv import column header e.g. ="-"}

\item{timepos}{integer: which position in delimited column header represents the retention time (csv only)}

\item{st}{numeric: sigma t - time similarity decay value}

\item{sr}{numeric: sigma r - correlational similarity decay value}

\item{maxt}{numeric: maximum time difference to calculate retention similarity for - all values beyond this are assigned similarity of zero}

\item{deepSplit}{logical: controls how agressively the HCA tree is cut - see ?cutreeDynamicTree}

\item{blocksize}{integer: number of features (scans?) processed in one block  =1000,}

\item{mult}{numeric: internal value, can be used to influence processing speed/ram usage}

\item{hmax}{numeric: precut the tree at this height, default 0.3 - see ?cutreeDynamicTree}

\item{sampNameCol}{integer: which column from the csv file contains sample names?}

\item{collapse}{logical: reduce feature intensities to spectrum intensities?}

\item{usePheno}{logical: tranfer phenotype data from XCMS object to SpecAbund dataset?}

\item{mspout}{logical: write msp formatted specta to file?}

\item{ExpDes}{either an R object created by R ExpDes object: data used for record keeping and labelling msp spectral output}

\item{normalize}{character: either "none", "TIC", or "quantile" normalization of feature intensities}

\item{minModuleSize}{integer: how many features must be part of a cluster to be returned? default = 2}

\item{linkage}{character: heirarchical clustering linkage method - see ?hclust}

\item{mzdec}{integer: number of decimal places used in printing m/z values}

\item{cor.method}{character: which correlational method used to calculate 'r' - see ?cor}
}
\value{
a ranclustR object.  base structure is that of a standard R heirarchical clustering output
}
\description{
Main clustering function
}
\details{
This is the Details section
}
\author{
Corey Broeckling
}
