% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.gcei.mat.R
\name{write.gcei.mat}
\alias{write.gcei.mat}
\title{write.gcei.mat}
\usage{
write.gcei.mat(ramclustObj = NULL, out.dir = NULL)
}
\arguments{
\item{ramclustObj}{ramclustR object to annotate.}

\item{out.dir}{valid directory path describing output directory/file location.}
}
\value{
nothing, just exports files to the working directory
}
\description{
Export GC-MS EI spectra for spectral searching in MSFinder
}
\details{
exports files to a directory called 'spectra'.  a new directory 'spectra/mat' is created to hold the individual mat files.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
\concept{clustering}
\concept{interpretMSSpectrum}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
